% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{get_power}
\alias{get_power}
\title{Calculate battery power}
\usage{
get_power(capacity, c_rate)
}
\arguments{
\item{capacity}{numeric scalar, the battery capacity.}

\item{c_rate}{numeric scalar, the battery charge rate.}
}
\value{
Returns a numeric scalar, the battery power.
}
\description{
\code{get_power} returns the battery power given capacity and charge
    rate.
}
\examples{
get_power(1, 2)
\dontrun{
    get_power("one", "two")
}
}
