% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spot_price_2016}
\alias{spot_price_2016}
\title{Electricity Market Spot Prices for April 2016 - April 2017}
\format{
A dataframe with 17558 rows and 4 variables:
\describe{
  \item{\code{timestamp}}{Date and time (UTC) at beginning of settlement
    each period.}
  \item{\code{N2EX}}{N2EX market spot price for settlement period, £/MWh.}
  \item{\code{EPEX}}{EPEX market spot price for settlement period, £/MWh.}
  \item{\code{SSP}}{SSP market spot price for settlement period, £/MWh.}
}
}
\usage{
spot_price_2016
}
\description{
A dataset containing a year of electricity spot prices for the N2EX, EPEX and
SSP markets.
}
\keyword{datasets}
