% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spot_price_au_2017}
\alias{spot_price_au_2017}
\title{Australian Electricity Market Spot Prices for July 2017 - June 2018}
\format{
A dataframe with 17520 rows and 2 variables:
\describe{
  \item{\code{timestamp}}{Local date and time at beginning of each
  settlement period.}
  \item{\code{price}}{Spot price for settlement period, £/MWh.}
}
}
\usage{
spot_price_au_2017
}
\description{
A dataset containing a year of electricity spot prices for the Australian
market.
}
\keyword{datasets}
