% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spot_price_uk_2018}
\alias{spot_price_uk_2018}
\title{UK Electricity Market Spot Prices for October 2018 - September 2019}
\format{
A dataframe with 17520 rows and 2 variables:
\describe{
  \item{\code{timestamp}}{Local date and time at beginning of each
  settlement period.}
  \item{\code{price}}{Spot price for settlement period, £/MWh.}
}
}
\usage{
spot_price_uk_2018
}
\description{
A dataset containing a year of electricity spot prices for the UK market.
}
\keyword{datasets}
